\name{BrainCoVersion}
\alias{BrainCoVersion}
\title{BrainConductor version}
\description{

  This function reports the Bioconductor version in use, as determined
  by the BiocInstaller package.

}
\usage{
BrainCoVersion()
}

\value{
  \code{package_version} representing the Bioconductor version in use.
}
\seealso{
  \code{\link{BCoInstall}} Installs/updates BrainConductor/CRAN packages.
}
\examples{
BrainCoVersion()
}
\keyword{environment}
